arc: reset cache states on halt/resume
Currently cache states, which are indicating
if each cache is flushed/invalidated, are
reseting on init, step, single_step and poll operations.
But in case of halt and resume, this does not happens and
some data still remains in cache and is not flushed in memory.
With this commit we add arc32_reset_caches_states() calls
in halt and resume functions so we would get/set actual data.

Signed-off-by: Evgeniy Didin <didin@synopsys.com>